+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlaygroundSettings.swift
@@ -321,7 +321,7 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
 
         case automatic
         case never
-        case hidden
+        case walletButtonHidden
     }
 
     enum ForceOnelink: String, PickerEnum {