+++ Example/PaymentSheet Example/PaymentSheet Example/PlaygroundController.swift
@@ -14,7 +14,7 @@ import Contacts
 import PassKit
 @_spi(STP) import StripeCore
 @_spi(STP) import StripePayments
-@_spi(STP) @_spi(PaymentSheetSkipConfirmation) @_spi(ExperimentalAllowsRemovalOfLastSavedPaymentMethodAPI) @_spi(PaymentMethodOptionsSetupFutureUsagePreview) @_spi(CardFundingFilteringPrivatePreview) @_spi(LinkHiddenWalletButtonPreview) import StripePaymentSheet
+@_spi(STP) @_spi(PaymentSheetSkipConfirmation) @_spi(ExperimentalAllowsRemovalOfLastSavedPaymentMethodAPI) @_spi(PaymentMethodOptionsSetupFutureUsagePreview) @_spi(CardFundingFilteringPrivatePreview) import StripePaymentSheet
 import SwiftUI
 import UIKit
 
@@ -143,8 +143,8 @@ import UIKit
             return PaymentSheet.LinkConfiguration(display: .automatic, brand: brand)
         case .never:
             return PaymentSheet.LinkConfiguration(display: .never, brand: brand)
-        case .hidden:
-            return PaymentSheet.LinkConfiguration(display: .hidden, brand: brand)
+        case .walletButtonHidden:
+            return PaymentSheet.LinkConfiguration(display: .walletButtonHidden, brand: brand)
         }
     }
     var customerConfiguration: PaymentSheet.CustomerConfiguration? {