+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentMethodAvailabilityTests.swift
@@ -7,7 +7,7 @@
 
 @testable @_spi(STP) import StripeCore
 @testable @_spi(STP) import StripePayments
-@testable @_spi(STP) @_spi(LinkHiddenWalletButtonPreview) import StripePaymentSheet
+@testable @_spi(STP) import StripePaymentSheet
 @testable @_spi(STP) import StripePaymentsTestUtils
 import XCTest
 
@@ -185,16 +185,16 @@ final class PaymentMethodAvailabilityTests: XCTestCase {
         XCTAssertFalse(isLinkEnabled, "Link should be disabled when display is set to .never")
     }
 
-    func testIsLinkEnabled_linkDisplayHidden_linkStillEnabled() {
+    func testIsLinkEnabled_linkDisplayWalletButtonHidden_linkStillEnabled() {
         let elementsSession = STPElementsSession._testValue(
             paymentMethodTypes: ["card"],
             isLinkPassthroughModeEnabled: true
         )
         var configuration = PaymentSheet.Configuration()
-        configuration.link = .init(display: .hidden)
+        configuration.link = .init(display: .walletButtonHidden)
         let isLinkEnabled = PaymentSheet.isLinkEnabled(elementsSession: elementsSession, configuration: configuration)
 
-        XCTAssertTrue(isLinkEnabled, "Link should remain enabled when display is set to .hidden")
+        XCTAssertTrue(isLinkEnabled, "Link should remain enabled when display is set to .walletButtonHidden")
     }
 
     func testShouldShowLinkButton_linkDisplayAutomatic_buttonShown() {
@@ -221,16 +221,16 @@ final class PaymentMethodAvailabilityTests: XCTestCase {
         XCTAssertFalse(shouldShowLinkButton, "Link button should not be shown when display is set to .never")
     }
 
-    func testShouldShowLinkButton_linkDisplayHidden_buttonNotShown() {
+    func testShouldShowLinkButton_linkDisplayWalletButtonHidden_buttonNotShown() {
         let elementsSession = STPElementsSession._testValue(
             paymentMethodTypes: ["card"],
             isLinkPassthroughModeEnabled: true
         )
         var configuration = PaymentSheet.Configuration()
-        configuration.link = .init(display: .hidden)
+        configuration.link = .init(display: .walletButtonHidden)
 
-        XCTAssertTrue(PaymentSheet.isLinkEnabled(elementsSession: elementsSession, configuration: configuration), "Link should remain enabled when display is set to .hidden")
-        XCTAssertFalse(PaymentSheet.shouldShowLinkButton(elementsSession: elementsSession, configuration: configuration), "Link button should not be shown when display is set to .hidden, even though Link remains enabled")
+        XCTAssertTrue(PaymentSheet.isLinkEnabled(elementsSession: elementsSession, configuration: configuration), "Link should remain enabled when display is set to .walletButtonHidden")
+        XCTAssertFalse(PaymentSheet.shouldShowLinkButton(elementsSession: elementsSession, configuration: configuration), "Link button should not be shown when display is set to .walletButtonHidden, even though Link remains enabled")
     }
 
     func testIsLinkEnabled_automaticTaxBilling_linkDisabled() {